App class 분리
